requirements.txt
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2019.9.11
chardet==3.0.4
click==6.7
click-log==0.1.8
config-templater==0.1.1
Django==3.0.6
django-background-tasks==1.2.5
django-bootstrap4==1.1.1
django-compat==1.0.15
djangorestframework==3.11.0
dnacentersdk==1.3.0.post7
fastjsonschema==2.14.4
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
immutables==0.14
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.15.1
MarkupSafe==1.1.1
nltk==3.5
pyadaptivecards==0.1.0
PyJWT==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
regex==2020.6.8
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
sklearn-crfsuite==0.3.6
smartsheet-python-sdk==2.86.0
soupsieve==2.0
sqlparse==0.3.1
tabulate==0.8.7
tqdm==4.46.1
urllib3==1.25.7
webexteamssdk==1.3
webhooksimple==0.1.2
Werkzeug==0.16.0
whitenoise==5.0.1
XlsxWriter==1.2.8
